Refactor - push EC2 host discovery into seperate package; not an elastic concern even if elastic is the only user.
