+++ media-api/app/lib/MediaApiConfig.scala
@@ -2,7 +2,7 @@ package lib
 
 import com.amazonaws.services.ec2.{AmazonEC2, AmazonEC2ClientBuilder}
 import com.gu.mediaservice.lib.config.CommonConfig
-import com.gu.mediaservice.lib.elasticsearch.EC2._
+import com.gu.mediaservice.lib.discovery.EC2._
 import play.api.Configuration
 
 import scala.util.Try
